LPS-121300 only delete expando tables related to the specific instance where the import is ocurring

+++ modules/apps/expando/expando-export-import/src/main/java/com/liferay/expando/exportimport/internal/staged/model/repository/StagedExpandoTableStagedModelRepository.java
@@ -87,8 +87,12 @@ public class StagedExpandoTableStagedModelRepository
 			_expandoTableLocalService.getExpandoTables(
 				QueryUtil.ALL_POS, QueryUtil.ALL_POS);
 
+		long companyId = portletDataContext.getCompanyId();
+
 		for (ExpandoTable expandoTable : expandoTables) {
-			_expandoTableLocalService.deleteTable(expandoTable);
+			if (expandoTable.getCompanyId() == companyId) {
+				_expandoTableLocalService.deleteTable(expandoTable);
+			}
 		}
 	}
